+++ components/modulos/shader/shader.h
@@ -13,20 +13,20 @@ namespace graphics{
 	public:
 	
 		std::vector<Shader*> shadersData; //Contiene un vector de código para cada shader
-		std::vector<short int> shadersCompiled;
-		std::vector<short int> shadersId;
-		std::map<const char*, short int> shadersPathId;
+		std::vector<unsigned int> shadersCompiled;
+		std::vector<unsigned int> shadersId;
+		std::map<const char*, unsigned int> shadersPathId;
 		//std::map<int, std::tuple<std::vector<std::tuple<Shader*, Graphics::Shader::TYPE_SHADER>>, unsigned int, std::vector<short int>>> shadersStack;
-		std::map<int, std::tuple<std::vector<Graphics::Shader::TYPE_SHADER>, unsigned int, std::vector<short int>>> shadersStack;
+		std::map<int, std::tuple<std::vector<Graphics::Shader::TYPE_SHADER>, unsigned int, std::vector<unsigned int>>> shadersStack;
 
 	private:
-		short int idS;
+		unsigned int idS;
 		Graphics::Shader::TYPE_SHADER tipo;
 		byte* codigo = 0;
 		int longitudCodigo;
 		bool compilado = false;
-		void cargar(short int id, Graphics::Shader::TYPE_SHADER type, const byte*& data, int count);
-		void stackShader(Shader* s, short int id, Graphics::Shader::TYPE_SHADER type);
+		void cargar(unsigned int id, Graphics::Shader::TYPE_SHADER type, const byte*& data, int count);
+		void stackShader(Shader* s, unsigned int id, Graphics::Shader::TYPE_SHADER type);
 	public:
 		Shader(Entity* entity, modules::graphics::Graphic* g, Component* p = NULL);
 		~Shader();
@@ -35,29 +35,29 @@ namespace graphics{
 		T* getData();
 
 		Graphics::Shader::TYPE_SHADER getType();
-		short int getId();
+		unsigned int getId();
 
 		int getShaderSize();
-		const Shader* getShader(int id);
+		const Shader* getShader(unsigned int id);
 	
-		int loadShader(const char* path, Graphics::Shader::TYPE_SHADER type);
-		void reloadShader(const char* path, Graphics::Shader::TYPE_SHADER type, int idShader, int idProgram);
-		int compileShader();
-		std::vector<short>compileShaders();
-		int compileShader(int id);
-		int addShader(const char * path, unsigned int id, Graphics::Shader::TYPE_SHADER type);
-		int addShader(const char* path, unsigned int id, Graphics::Shader::TYPE_SHADER type, const byte* data, int count);
-		int addShader(const char* path, unsigned int id, Graphics::Shader::TYPE_SHADER type, const char* data, int count);
-		int addShader(unsigned int id, Graphics::Shader::TYPE_SHADER type);
-		int addShader(unsigned int id, Graphics::Shader::TYPE_SHADER type, const byte* data, int count);
-		int addShader(unsigned int id, Graphics::Shader::TYPE_SHADER type, const char* data, int count);
+		unsigned int loadShader(const char* path, Graphics::Shader::TYPE_SHADER type);
+		void reloadShader(const char* path, Graphics::Shader::TYPE_SHADER type, unsigned int idShader, unsigned int idProgram);
+		unsigned int compileShader();
+		std::vector<unsigned int>compileShaders();
+		unsigned int compileShader(unsigned int id);
+		unsigned int addShader(const char * path, unsigned int id, Graphics::Shader::TYPE_SHADER type);
+		unsigned int addShader(const char* path, unsigned int id, Graphics::Shader::TYPE_SHADER type, const byte* data, int count);
+		unsigned int addShader(const char* path, unsigned int id, Graphics::Shader::TYPE_SHADER type, const char* data, int count);
+		unsigned int addShader(unsigned int id, Graphics::Shader::TYPE_SHADER type);
+		unsigned int addShader(unsigned int id, Graphics::Shader::TYPE_SHADER type, const byte* data, int count);
+		unsigned int addShader(unsigned int id, Graphics::Shader::TYPE_SHADER type, const char* data, int count);
 
 		
 		/**/
 		std::vector<Shader*>* getShaders();
 		//Para agilizar mantendremos un vector para los compilados
-		std::vector<short int>* getShadersCompiled();
-		std::vector<short>* getShadersId(int idStack=0);
+		std::vector<unsigned int>* getShadersCompiled();
+		std::vector<unsigned int>* getShadersId(int idStack=0);
 		bool isCompiled() { return compilado; }
 	};
 	template<class T>
